refactor: go to tsx and add customer wrapper on form
