+++ packages/apps/intervention/src/utils/form-types.ts
@@ -1,0 +1,9 @@
+export interface CustomComponentProps {
+  style?: any;
+  title?: string;
+  defaultValue?: string;
+  onChange?: (any: any) => void;
+  readonly?: boolean;
+  required?: boolean;
+  [key: string]: any;
+}
